Update Mobile UI
This update modifies the layout slightly to swap current_image and batch_images. This allows current_images to be rendered at the top in a single column layout.

It introduces a mobile script to handle mobile interactions and detection

Updates the associated scripts to handle correct sizing and eventing on mobile screens

Update the css stylesheets to account for smaller screen sizes