Add 16bits color depth support
Pillow doesn't support natively so OpenCV was used when processing images requested as 16bits.
That image is later converted to a Pillow image when passing it to the comfy ProgressBar.
When requesting an image as 8bits RGB (the default for all formats), or when requesting a format other than PNG, the original image processing is used.