fix: some gen params not reset correctly when preset modal initializes
When the Create/Edit Preset modal is opened, `clearPresetView` loops over gen parameters to set them to the appropriate default states.  It was doing this via `getPresetTypes('input')` which loops over enabled Input controls (as in, those rendered in the left sidebar's Inputs pane).  But we actually want to loop over enabled Preset Input controls (as in, those rendered within the Create/Edit Preset modal).