image browser performance improvement - set initial aspect ratio of thumbnails
In most cases, we know the aspect ratio of the image we will be using as
the thumbnail.  By setting this aspect ratio when creating the thumbnail
element, we can avoid the reflow that occurs when the image loads.

In my test where I click Refresh while viewing folder c (in the a/b/c tree)
where c has 635 files and I have the profiler running (which slows things down)
this reduces the time from click to finish by about 20% (from 8s to 6.5s).  With debugger closed, the click-to-finish time is about 5s.