Item browser performance improvements (#908)
* image browser performance improvement - set initial aspect ratio of thumbnails

In most cases, we know the aspect ratio of the image we will be using as
the thumbnail.  By setting this aspect ratio when creating the thumbnail
element, we can avoid the reflow that occurs when the image loads.

In my test where I click Refresh while viewing folder c (in the a/b/c tree)
where c has 635 files and I have the profiler running (which slows things down)
this reduces the time from click to finish by about 20% (from 8s to 6.5s).  With debugger closed, the click-to-finish time is about 5s.

* Avoid repetitive reflow when triggering lazyload elements

* Defer adding title to an item until the mouse enters the item.

Converting HTML to text is expensive, and 99% of the time not needed.

* only set scrollOffset if the user had been scrolled.

* Optimize the fetching and updating during initial load and on refresh.

* Only render the files once after the folder tree has been fetched and expanded (original code was rendering files each fetch...so could be many many times)
* Only fetch the folders/files once if refreshing the root folder (original code was fetching 2x)
* Only fetch the folders/files twice if refreshing a subfolder (original code was fetching (1 + 1 * subfolderDepth).

* clean and simplify

* remove stray css edits

---------

Co-authored-by: Alex "mcmonkey" Goodwin <git_commits@alexgoodwin.dev>