Add live previews to the Full View
In order to support live previews, the Full View must now keep track of the batch ID, so the function `ImageFullViewHelper.showImage` has a new parameter to pass in the batch ID.  This is saved as the `this.currentBatchId` member.
The `img` element with id `imageview_popup_modal_img` is also exposed as `this.imgElement`.

With the img element exposed, and the Batch ID now associated with the Full View, the function `GenerateHandler.internalHandleData` can now safely update the Full View's img src when its batch ID matches.