Enlarge Plus Sign Gen Button (#1204)
* Enlarge Plus Sign Gen Button

Edited some of the genpage css to make the plus sign button to the left of the prompt box look bigger and more noticeable. Added margin-right to the textbox itself to ensure it doesn't go behind the Generate Button.

* Update genpage.css

Removed a few unnecessary changes to the css code and made the + button slightly smaller.

* simpler css align

---------

Co-authored-by: Alex "mcmonkey" Goodwin <git_commits@alexgoodwin.dev>